+++ CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added polygon layer to map
 - Polygon layer queried with slider
 - Added basemaps
+- Antecedant moisture selector
 
 ### Changed
 
 - Updated README with contributors and project info
 - Changed favicon to WIM favicon
+- Restyled sidebar
 
 
 ### Fixed